Relax phoenix constraint version and use unreleased bamboo
What changed?
=============

We relax the version of Phoenix that can be used since that could cause
problems for libraries depending on `bamboo_phoenix`.

We also depend on the unreleased version of Bamboo that removes the
`Bamboo.Phoenix` module so that we don't get errors when trying to
compile because we're redefining the module.