Set bamboo dep to 1.7.1 or greater
What changed?
==============

We set the bamboo dependency to >= 1.7.1. This is technically
inaccurate, since Bamboo v1.7.1 still has the `Bamboo.Phoenix` module
defined. Thus, an app including both `bamboo` and `bamboo_phoenix` that
compiles with `--warnings-as-errors` will be unable to compile due to
the module being overwritten.

So why do this?
==============

We cannot publish a library to hex with a `git` dependency. From the
docs:

> Only Hex packages may be used as dependencies of the package. It is
not possible to upload packages with Git dependencies.

So we want to publish a technically incorrect version since it's version
`0.1.0`, and when we change to `Bamboo` 2.0, we can publish version
`1.0` of `bamboo_phoenix`.