Remove --warnings-as-errors from CI build
What changed?
=============

We remove the `--warnings-as-errors` flag from `mix compile` in CI for
now. We know the CI will fail because we are currently redefining the
`Bamboo.Phoenix` module in our `bamboo` dependency. That will be fixed
when we can depend on bamboo 2.0, and then we can reintroduce the
`--warnings-as-errors` in CI.